Cortex Filer Handler: Allow caller to specify container name, get short-lived URLs and be able to delete files by hash
* Cortex Filer Handler: Allow caller to specify bucket name

* Add ability to create short-lived URLs

* Read container name in blobHandler

* Fix race condition when reading container name

* Fix failing tests

* Fix some test failures

* Update StorageService.js

* Update StorageFactory.js

* Add delete file by hash

* Update test-azure.sh

* Revert test changes

* Introduce mock redis and fix some test issues

* Add tests for container name parsing

* Add tests for delete file by hash

* Add test for checkHash returning short-lived URLs

+++ helper-apps/cortex-file-handler/tests/conversionResilience.test.js
@@ -8,7 +8,7 @@ import XLSX from "xlsx";
 import { port } from "../src/start.js";
 import { gcs, GCS_BUCKETNAME } from "../src/blobHandler.js";
 import { getFileStoreMap, setFileStoreMap } from "../src/redis.js";
-import { cleanupHashAndFile } from "./testUtils.helper.js";
+import { cleanupHashAndFile, startTestServer, stopTestServer } from "./testUtils.helper.js";
 import { gcsUrlExists } from "../src/blobHandler.js";
 
 const baseUrl = `http://localhost:${port}/api/CortexFileHandler`;
@@ -47,13 +47,18 @@ function isGcsConfigured() {
   return !!gcs;
 }
 
+// Ensure server is ready before tests
 test.before(async (t) => {
+  await startTestServer();
   const dir = path.join(fs.mkdtempSync(path.join(process.cwd(), "conv-test-")));
   t.context.tmpDir = dir;
 });
 
 test.after.always(async (t) => {
-  fs.rmSync(t.context.tmpDir, { recursive: true, force: true });
+  await stopTestServer();
+  if (t.context?.tmpDir) {
+    fs.rmSync(t.context.tmpDir, { recursive: true, force: true });
+  }
 });
 
 // 1. Remote-URL upload path should still return converted info
